Add an alternative to Duration inside Sound sprite. (EndIndex)

+++ src/misc/save_file.cpp
@@ -16,7 +16,7 @@ bool			bunny_save_file(const char		*file,
 
   if ((fd = open(file, O_CREAT | O_TRUNC | O_WRONLY | O_BINARY, 0644)) == -1)
     scream_error_if(return (false), bunny_errno, PATTERN, "ressource,file", file, data, len, "false");
-  if (write(fd, data, len) != len)
+  if (write(fd, data, len) != (ssize_t)len)
     {
       close(fd);
       scream_error_if(return (false), bunny_errno, PATTERN, "ressource,file", file, data, len, "false");
